fix(ci): UI 测试首跑 3 失败——锁屏容器标识覆盖 + E2E 步进 + Tab 文案 + SOS 0.6s 提示
CI 34021989599（L1 UI 测试首次真实执行暴露）：

1) LockOverlayView 容器 .accessibilityIdentifier 未配 accessibilityElement
   (children: .contain)，SwiftUI 把容器标识下放覆盖子元素——解锁按钮的
   SP-01.lockOverlay.unlock 被顶成 SP-01.lockOverlay（层级 dump 实证），
   锁屏用例找不到解锁按钮。
2) E2E 切片：V3.72 六步编排在建档与拍摄之间新增 FR21.9 ④ 添加家人步，
   测试未步进即找 SP-07——补 skip 步进。
3) 五 Tab 文案随 L10n 重写更新（记录→健康档案、AI→AI 助手），测试期望
   同步当前译文。
4) 层级 dump 连带暴露 SOS 无障碍提示「长按 0 秒激活」：Int(0.6) 截断 +
   %1$d 整数格式——改 Double + %1$.1f（三语同步），§7.1 防误触语义在
   提示层恢复真实时长。

L0 14/14 全绿。

+++ App/Localization/L10n.swift
@@ -2243,7 +2243,9 @@ enum L10n {
     static func f19MarkTakenNoMatch(_ name: String) -> String { String(format: t("f19.markTakenNoMatchFmt"), name) }
     static func f19MarkTakenDone(_ name: String) -> String { String(format: t("f19.markTakenDoneFmt"), name) }
     static func f19MarkTakenMultiple(_ names: String) -> String { String(format: t("f19.markTakenMultipleFmt"), names) }
-    static func emergency_sos_holdA11y(_ seconds: Int) -> String {
+    static func emergency_sos_holdA11y(_ seconds: Double) -> String {
+        // 必须 Double + %1$.1f：Int 截断把 0.6 秒显示成「长按 0 秒激活」
+        // （CI 34021989599 层级 dump 实证）——SOS 防误触语义在提示层失真
         String(format: t("emergency.sosHoldA11yFmt"), seconds)
     }
     static var helpcard_photoPending: String { t("helpcard.photoPending") }
